Move api views
