+++ app/experimenter/experiments/api/v1/views.py
@@ -1,0 +1,35 @@
+from rest_framework.generics import (
+    ListAPIView,
+    RetrieveAPIView,
+)
+
+
+from experimenter.experiments.constants import ExperimentConstants
+from experimenter.experiments.models import Experiment
+from experimenter.experiments.serializers.entities import ExperimentSerializer
+from experimenter.normandy.serializers import ExperimentRecipeSerializer
+
+
+class ExperimentListView(ListAPIView):
+    filter_fields = ("status",)
+    queryset = Experiment.objects.all()
+    serializer_class = ExperimentSerializer
+
+
+class ExperimentDetailView(RetrieveAPIView):
+    lookup_field = "slug"
+    queryset = Experiment.objects.all()
+    serializer_class = ExperimentSerializer
+
+
+class ExperimentRecipeView(RetrieveAPIView):
+    lookup_field = "slug"
+    queryset = Experiment.objects.filter(
+        status__in=(
+            ExperimentConstants.STATUS_SHIP,
+            ExperimentConstants.STATUS_ACCEPTED,
+            ExperimentConstants.STATUS_LIVE,
+            ExperimentConstants.STATUS_COMPLETE,
+        )
+    )
+    serializer_class = ExperimentRecipeSerializer
