+++ app/experimenter/experiments/api/v2/views.py
@@ -28,31 +28,6 @@ from experimenter.experiments.serializers.timeline_population import (
 from experimenter.normandy.serializers import ExperimentRecipeSerializer
 
 
-class ExperimentListView(ListAPIView):
-    filter_fields = ("status",)
-    queryset = Experiment.objects.all()
-    serializer_class = ExperimentSerializer
-
-
-class ExperimentDetailView(RetrieveAPIView):
-    lookup_field = "slug"
-    queryset = Experiment.objects.all()
-    serializer_class = ExperimentSerializer
-
-
-class ExperimentRecipeView(RetrieveAPIView):
-    lookup_field = "slug"
-    queryset = Experiment.objects.filter(
-        status__in=(
-            ExperimentConstants.STATUS_SHIP,
-            ExperimentConstants.STATUS_ACCEPTED,
-            ExperimentConstants.STATUS_LIVE,
-            ExperimentConstants.STATUS_COMPLETE,
-        )
-    )
-    serializer_class = ExperimentRecipeSerializer
-
-
 class ExperimentSendIntentToShipEmailView(UpdateAPIView):
     lookup_field = "slug"
     queryset = Experiment.objects.filter(status=Experiment.STATUS_REVIEW)
