+++ app/experimenter/urls.py
@@ -7,10 +7,8 @@ from experimenter.experiments.views import ExperimentListView
 
 
 urlpatterns = [
-    re_path(r"^api/v1/experiments/", include("experimenter.experiments.public_api_urls")),
-    re_path(
-        r"^api/v2/experiments/", include("experimenter.experiments.private_api_urls")
-    ),
+    re_path(r"^api/v1/experiments/", include("experimenter.experiments.api.v1.urls")),
+    re_path(r"^api/v2/experiments/", include("experimenter.experiments.api.v2.urls")),
     re_path(r"^admin/", admin.site.urls),
     re_path(r"^experiments/", include("experimenter.experiments.web_urls")),
     re_path(r"^$", ExperimentListView.as_view(), name="home"),
